修复上传APK，添加任务调度相关的BUG

+++ whalegoods-erp/whalegoods-erp/src/main/java/com/whalegoods/mapper/DeviceRoadMapper.java
@@ -21,6 +21,6 @@ public interface DeviceRoadMapper extends BaseMapper<DeviceRoad,String>  {
     
     List<DeviceRoad> selectListByDeviceIdForExcel(String deviceId);
     
-    List<Integer> selectExistPrice(DeviceRoad objCdt);
+    List<Double> selectExistPrice(DeviceRoad objCdt);
     
 }
